requirements.txt
customtkinter==4.6.3
Pillow==9.3.0
